+++ cg/services/orders/store_order_services/store_fastq_order_service.py
@@ -1,14 +1,9 @@
 import logging
 from datetime import datetime
 
-from cg.constants import (
-    DataDelivery,
-    GenePanelMasterList,
-    Priority,
-    SeqLibraryPrepCategory,
-    Workflow,
-)
+from cg.constants import DataDelivery, GenePanelMasterList, Priority, Workflow
 from cg.constants.constants import CustomerId
+from cg.constants.sequencing import SeqLibraryPrepCategory
 from cg.exc import OrderError
 from cg.models.orders.order import OrderIn
 from cg.models.orders.sample_base import StatusEnum
