+++ tests/fixture_plugins/quality_controller_fixtures/sequencing_qc_check_scenario.py
@@ -1,6 +1,6 @@
-from cg.constants import SeqLibraryPrepCategory
 from cg.constants.constants import Workflow
 from cg.constants.priority import Priority
+from cg.constants.sequencing import SeqLibraryPrepCategory
 from cg.store.models import Application, ApplicationVersion, Case, Sample
 from cg.store.store import Store
 from tests.conftest import StoreHelpers
