+++ tests/fixture_plugins/quality_controller_fixtures/sequencing_qc_fixtures.py
@@ -1,8 +1,8 @@
 import pytest
 
-from cg.constants import SeqLibraryPrepCategory
 from cg.constants.constants import Workflow
 from cg.constants.priority import Priority
+from cg.constants.sequencing import SeqLibraryPrepCategory
 from cg.store.models import Case, Sample
 from tests.fixture_plugins.quality_controller_fixtures.sequencing_qc_check_scenario import (
     SequencingQCCheckScenarios,
