+++ tests/store/crud/conftest.py
@@ -3,8 +3,9 @@ from typing import Generator
 
 import pytest
 
-from cg.constants import SeqLibraryPrepCategory, Workflow
+from cg.constants import Workflow
 from cg.constants.constants import CustomerId
+from cg.constants.sequencing import SeqLibraryPrepCategory
 from cg.constants.subject import PhenotypeStatus
 from cg.store.models import Case, CaseSample, Customer, Order, Sample
 from cg.store.store import Store
